left graph-update, still there are errors

+++ src/view/graph/Mygraph.java
@@ -1,9 +1,6 @@
 package view.graph;
 
-import javafx.beans.property.DoubleProperty;
-import javafx.beans.property.FloatProperty;
-import javafx.beans.property.SimpleDoubleProperty;
-import javafx.beans.property.SimpleFloatProperty;
+import javafx.beans.property.*;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.layout.AnchorPane;
@@ -14,7 +11,7 @@ import java.io.IOException;
 
 public class Mygraph extends Pane {
     public FloatProperty listvalue;
-    public DoubleProperty time;
+    public IntegerProperty time;
     public Mygraph() {
         super();
 
@@ -23,7 +20,7 @@ public class Mygraph extends Pane {
             Parent root = fxl.load(getClass().getResource("Mygraph.fxml").openStream());
             Mygraphcontroller mygraphcontroller = fxl.getController();
             this.listvalue=new SimpleFloatProperty();
-            this.time= new SimpleDoubleProperty();
+            this.time= new SimpleIntegerProperty();
             mygraphcontroller.listvalue.bind(this.listvalue);
             mygraphcontroller.time.bind(this.time);
             this.time.addListener((o,ov,nv)->mygraphcontroller.AddtoGraph());
